fix: extract Machine class_methods to it own file to reduce file loc
